+++ 07-spring-boot-rest-crud-employee/00-spring-boot-spring-mvc-crud-starter-code/src/main/java/np/com/krishnabk/springboot/thymeleafdemo/Application.java
@@ -1,4 +1,4 @@
-package np.com.krishnabk.employeecrud;
+package np.com.krishnabk.springboot.thymeleafdemo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
