+++ 07-spring-boot-rest-crud-employee/00-spring-boot-spring-mvc-crud-starter-code/src/main/java/np/com/krishnabk/springboot/thymeleafdemo/entity/Employee.java
@@ -1,4 +1,4 @@
-package np.com.krishnabk.employeecrud.entity;
+package np.com.krishnabk.springboot.thymeleafdemo.entity;
 
 import jakarta.persistence.*;
