+++ 07-spring-boot-rest-crud-employee/00-spring-boot-spring-mvc-crud-starter-code/src/main/java/np/com/krishnabk/springboot/thymeleafdemo/service/EmployeeServiceImpl.java
@@ -1,10 +1,9 @@
-package np.com.krishnabk.employeecrud.service;
+package np.com.krishnabk.springboot.thymeleafdemo.service;
 
-import np.com.krishnabk.employeecrud.dao.EmployeeRepository;
-import np.com.krishnabk.employeecrud.entity.Employee;
+import np.com.krishnabk.springboot.thymeleafdemo.dao.EmployeeRepository;
+import np.com.krishnabk.springboot.thymeleafdemo.entity.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
