+++ 07-spring-boot-rest-crud-employee/00-spring-boot-spring-mvc-crud-starter-code/src/test/java/np/com/krishnabk/springboot/thymeleafdemo/ApplicationTests.java
@@ -1,4 +1,4 @@
-package np.com.krishnabk.employeecrud;
+package np.com.krishnabk.springboot.thymeleafdemo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
